ci: updated pysindy

+++ Code/Methods/Codebase/Utils/utils_sindy.py
@@ -69,8 +69,8 @@ def savitzky_golay(y, window_size, order, deriv=0, rate=1):
     from math import factorial
 
     try:
-        window_size = np.abs(np.int(window_size))
-        order = np.abs(np.int(order))
+        window_size = np.abs(np.int_(window_size))
+        order = np.abs(np.int_(order))
     except ValueError:
         raise ValueError("window_size and order have to be of type int")
     if window_size % 2 != 1 or window_size < 1:
